pyproject.toml
[tool.poetry]
name = "simple-sql-lineage"
version = "0.1.0"
description = "Trying to make my life easier in some messed projects"
authors = ["Thom Moreira <sevencagedtigers@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
rich = "^13.6.0"
black = "^23.9.1"
scipy = "^1.11.3"
pyqt5 = "^5.15.9"
neo4j = "^5.13.0"
python-dotenv = "^1.0.0"
sqllineage = "1.4.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
